refac: 로직 역할 분리 작업
- hooks에서 createRoot, rerender 로직 분리
- hooks 글로벌에서 root 관련 로직 분리 하여 root파일에 작성
