--- src/components/child.tsx
@@ -1,3 +1,0 @@
-export default function Child({ name }: { name?: string }) {
-  return <p>child{name && `: ${name}`}</p>;
-}
