--- src/components/counter.tsx
@@ -1,12 +1,0 @@
-import { useState } from "@/utils/core/hooks";
-
-export default function Counter() {
-  const [count, setCount] = useState(0);
-  return (
-    <div>
-      <p>Count: {count}</p>
-      <button onClick={() => setCount(count + 1)}>+</button>
-      <button onClick={() => setCount(count - 1)}>-</button>
-    </div>
-  );
-}
