--- src/components/header.tsx
@@ -1,3 +1,0 @@
-export default function Header({ title }: { title: string }) {
-  return <header>{title}</header>;
-}
